Common.rs rewritten
